refactor: move popup selectors into account slice
